+++ Sources/SuperwallKit/Paywall/Presentation/PaywallInfo.swift
@@ -131,6 +131,10 @@ public final class PaywallInfo: NSObject {
   /// The state of the paywall, updated on paywall did dismiss.
   public let state: [String: Any]
 
+  /// Indicates how intro offer eligiblity should be treat on products. Defaults to
+  /// `.automatic`.
+  public let introOfferEligibility: IntroOfferEligibility
+
   init(
     databaseId: String,
     identifier: String,
@@ -163,7 +167,8 @@ public final class PaywallInfo: NSObject {
     surveys: [Survey],
     presentation: PaywallPresentationInfo,
     isScrollEnabled: Bool,
-    state: [String: Any]
+    state: [String: Any],
+    introOfferEligibility: IntroOfferEligibility
   ) {
     self.databaseId = databaseId
     self.identifier = identifier
@@ -231,6 +236,7 @@ public final class PaywallInfo: NSObject {
     self.closeReason = closeReason
     self.isScrollEnabled = isScrollEnabled
     self.state = state
+    self.introOfferEligibility = introOfferEligibility
   }
 
   func placementParams(
@@ -264,6 +270,7 @@ public final class PaywallInfo: NSObject {
       "build_id": buildId,
       "close_reason": closeReason.description,
       "is_scroll_enabled": isScrollEnabled as Any,
+      "intro_offer_eligibility": introOfferEligibility.description,
       "app_transaction_id": ReceiptManager.appTransactionId as Any
     ]
 
@@ -375,7 +382,8 @@ extension PaywallInfo: Stubbable {
         delay: 0
       ),
       isScrollEnabled: true,
-      state: [:]
+      state: [:],
+      introOfferEligibility: .automatic
     )
   }
 
@@ -424,7 +432,8 @@ extension PaywallInfo: Stubbable {
         delay: 0
       ),
       isScrollEnabled: true,
-      state: [:]
+      state: [:],
+      introOfferEligibility: .automatic
     )
   }
 }