+++ Sources/SuperwallKit/Paywall/View Controller/Web View/Message Handling/PaywallMessage.swift
@@ -68,10 +68,19 @@ enum PaywallMessage: Decodable, Equatable {
 
   case transactionRestore
   case transactionStart
-  case transactionComplete
+  case transactionComplete(trialEndDate: Date?, productIdentifier: String)
   case transactionFail
   case transactionAbandon
   case transactionTimeout
+
+  // swiftlint:disable:next enum_case_associated_values_count
+  case scheduleNotification(
+    type: LocalNotificationType,
+    title: String,
+    subtitle: String?,
+    body: String,
+    delay: Milliseconds
+  )
 
   private enum MessageTypes: String, Decodable {
     case onReady = "ping"
@@ -86,6 +95,7 @@ enum PaywallMessage: Decodable, Equatable {
     case userAttributesUpdated = "user_attribute_updated"
     case initiateWebCheckout = "initiate_web_checkout"
     case requestStoreReview = "request_store_review"
+    case scheduleNotification = "schedule_notification"
   }
 
   // Everyone write to eventName, other may use the remaining keys
@@ -102,13 +112,18 @@ enum PaywallMessage: Decodable, Equatable {
     case reviewType
     case browserType
     case checkoutContextId
+    case type
+    case title
+    case subtitle
+    case body
+    case delay
   }
 
   enum PaywallMessageError: Error {
     case decoding(String)
   }
 
-  // swiftlint:disable:next function_body_length
+  // swiftlint:disable:next function_body_length cyclomatic_complexity
   init(from decoder: Decoder) throws {
     let values = try decoder.container(keyedBy: CodingKeys.self)
     if let messageType = try? values.decode(MessageTypes.self, forKey: .messageType) {
@@ -176,6 +191,21 @@ enum PaywallMessage: Decodable, Equatable {
       case .requestStoreReview:
         if let reviewType = try? values.decode(ReviewType.self, forKey: .reviewType) {
           self = .requestStoreReview(reviewType)
+          return
+        }
+      case .scheduleNotification:
+        if let type = try? values.decode(LocalNotificationType.self, forKey: .type),
+          let title = try? values.decode(String.self, forKey: .title),
+          let body = try? values.decode(String.self, forKey: .body),
+          let delay = try? values.decode(Milliseconds.self, forKey: .delay) {
+          let subtitle = try values.decodeIfPresent(String.self, forKey: .subtitle)
+          self = .scheduleNotification(
+            type: type,
+            title: title,
+            subtitle: subtitle,
+            body: body,
+            delay: delay
+          )
           return
         }
       }