+++ Sources/SuperwallKit/Paywall/View Controller/Web View/Message Handling/PaywallWebEvent.swift
@@ -16,5 +16,6 @@ enum PaywallWebEvent: Equatable {
   case openedUrlInSafari(_ url: URL)
   case openedDeepLink(url: URL)
   case customPlacement(name: String, params: JSON)
+  case scheduleNotification(notification: LocalNotification)
   case userAttributesUpdated(attributes: JSON)
 }