+++ Sources/SuperwallKit/StoreKit/Transactions/TransactionManager.swift
@@ -245,15 +245,13 @@ final class TransactionManager {
             title: hasEntitlements ? "Restore via the web?" : "No Subscription Found",
             message: hasEntitlements ? hasSubsText : noSubsText,
             actionTitle: "Yes",
-            closeActionTitle: "Cancel",
-            // swiftlint:disable:next trailing_closure
-            action: {
-              guard let sharedApplication = UIApplication.sharedApplication else {
-                return
-              }
-              sharedApplication.open(restoreUrl)
+            closeActionTitle: "Cancel"
+          ) {
+            guard let sharedApplication = UIApplication.sharedApplication else {
+              return
             }
-          )
+            sharedApplication.open(restoreUrl)
+          }
           return .webRestore
         }
       }
@@ -894,11 +892,13 @@ final class TransactionManager {
 
     let paywallInfo: PaywallInfo
     let eventSource: InternalSuperwallEvent.Transaction.Source
+    let trialEndDate = product.trialPeriodEndDate
     switch source {
     case .internal(_, let paywallViewController):
       paywallInfo = await paywallViewController.info
       eventSource = .internal
-      await paywallViewController.webView.messageHandler.handle(.transactionComplete)
+      await paywallViewController.webView.messageHandler
+        .handle(.transactionComplete(trialEndDate: trialEndDate, productIdentifier: product.productIdentifier))
     case .purchaseFunc,
       .observeFunc:
       paywallInfo = .empty()
@@ -940,7 +940,11 @@ final class TransactionManager {
       let notifications = paywallInfo.localNotifications.filter {
         $0.type == .trialStarted
       }
-      await NotificationScheduler.scheduleNotifications(notifications, factory: factory)
+      await NotificationScheduler.shared.scheduleNotifications(
+        notifications,
+        fromPaywallId: paywallInfo.identifier,
+        factory: factory
+      )
     case .subscriptionStart:
       await Superwall.shared.track(
         InternalSuperwallEvent.SubscriptionStart(